change the amount in float

+++ database/migrations/2024_06_12_013041_create_payments.php
@@ -46,7 +46,7 @@ return new class extends Migration {
 
             // //Payment
             $table->string('payment_method');
-            $table->integer('amount');
+            $table->float('amount', 8, 2);
 
             $table->timestamps();
